Move ExistingSecretTestFixture from TestAwsSecret to module
As we will soon share it with TestConfig.
Renamed it ExistingAwsSecretTestFixture.
Cleaned it up and added minimal documentation.
